backend/pyproject.toml
[project]
name = "ezgo-poc-backend"
version = "0.1.0"
description = "ezGO POC Backend with FastAPI, SQLAlchemy, PostGIS, and H3"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.25",
    "psycopg2-binary==2.9.9",
    "geoalchemy2==0.14.3",
    "alembic==1.13.1",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "shapely==2.0.2",
    "python-dotenv==1.0.0",
    "h3==3.7.6",
    "numpy<2",
    "mapbox>=0.18.1",
    "ortools>=9.9.3963",
    "scikit-learn>=1.4.0",
    "hdbscan>=0.8.33",
    "scipy>=1.11.0",
    "requests>=2.31.0",
    "pyproj>=3.6.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

